Make reflector optional
